sphinx: add blinding key and shared secret derivation
This commit adds a BlindingKey field to our OnionPackets to allow
use of a blinding ephemeral key passed out of band for onions that
are part of a blinded route.

If a blinding key is included, shared secret generation switches out
our node's onion key for a blinded private key, which is derived using
a shared secret generated between our onion key and the blinding key
provided.

TODO:
- Do we need to persist this blinding key?
- SingleKeyECDH is extended here with Mul, is that safe?