multi: handle route blinding in ProcessOnionPacket
In this commit, ProcessOnionPacket is updated to take in a blinding key
as a parameter. This key is then used in order to determine the blinding
factor necessary to decrypt the onion. This change is accompanied with a
test that tests this change against a test vector from the route
blinding spec PR.